Fixed the bot to work with the latest version of discord.js.
Lots of work done polishing up individual commands, a few parts still aren't working quite right.

Outside the code, the program is now run with PM2 enabled, so if it crashes there should only be a blip of service. This shouldn't be a problem because very little information is stored in memory.